Auto-build JsonArrayBuilder or JsonObjectBuilders when used as JsonValues
In the builder patterns that I've used, using a builder inside another builder is fine and calling `build()` on the outer object automatically calls build on inner ones. This change would do the same thing for the `JsonArrayBuilder` and `JsonObjectBuilder`. It basically allows the two builders to be coerced into a `JsonValue` in the same way that a Collection or `Map` can be.